Read cTransformNode blocks as part of Scenegraph collections.
I assumed these cTransformNodes only existed as part of cResourceNode
but it looks like they can come up by themselves. Adding this now
allows all neighborhood decoration resources to be loaded.
