+++ Assets/Scripts/OpenTS2/Files/Formats/DBPF/Scenegraph/ScenegraphResourceCollection.cs
@@ -117,6 +117,7 @@ namespace OpenTS2.Files.Formats.DBPF.Scenegraph
                 { ResourceNodeBlock.TYPE_ID, new ResourceNodeBlockReader() },
                 { DataListExtensionBlock.TYPE_ID, new DataListExtensionBlockReader() },
                 { ShapeRefNodeBlock.TYPE_ID, new ShapeRefNodeBlockReader() },
+                { TransformNodeBlock.TYPE_ID, new TransformNodeBlockReader() },
             };
 
         private static ResourceKey ReadResourceKey(IoBuffer reader)
